Refactor models and DTOs to use Long for quantity fields; add new Solicitud and Sucursal repositories, and implement Detalle as a base class for better structure

+++ src/main/java/com/farmacia/dto/DetalleDTO.java
@@ -12,7 +12,7 @@ public class DetalleDTO {
     public static class GET{
         private Long id;
         private String producto;
-        private int cantidad;
+        private Long cantidad;
         private Float precio;
         private Float subtotal;
         private String lote;
